requirements.txt
alabaster==0.7.13
amqp==5.1.1
astropy==5.2.2
asttokens==2.2.1
async-timeout==4.0.2
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==4.1.0
casadata==2024.1.15
casatasks==6.5.5.21
casatools==6.5.5.21
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.6.23
codecov==2.1.13
comm==0.1.4
contourpy==1.1.0
coverage==7.2.7
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
docutils==0.18.1
exceptiongroup==1.1.2
executing==1.2.0
flower==2.0.0
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
h5py==3.8.0
humanize==4.7.0
idna==3.4
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
invoke==2.2.0
ipykernel==6.25.1
ipympl==0.9.3
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
kombu==5.3.1
lazy_loader==0.3
llvmlite==0.41.1
lz4==4.3.2
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mock==5.1.0
mypy==1.4.1
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numba==0.58.1
numpy==1.24.4
opencv-python==4.10.0.84
#-e git+http://git@github.com/ovro-lwa/distributed-pipeline.git@4ab4c281fe7ae47a18c6424993b75dee6496a47e#egg=ovro_lwa_orca
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
pockets==0.9.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyerfa==2.0.0.3
Pygments==2.16.1
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.14.0
python-casacore==3.5.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.6.0
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.13.2
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
widgetsnbextension==4.0.8
wrapt==1.15.0
zipp==3.16.2
reproject==0.13.1
multiprocess==0.70.18
torch==2.4.1
torchvision==0.19.1
git+https://github.com/facebookresearch/segment-anything.git
#git+https://github.com/SysCV/sam-hq.git
segment-anything-hq==0.3
timm==0.9.16
einops>=0.6
tqdm
reportlab
